fix show_board bug (stopped with added Xs and Os).
The problem was that I used the board's original elements (range of
integers 1..9) to track when a newline would be created. For example,
when the element 3 was reached, it would move from printing elements
inline to printing them on a new line. This works when the board is
populated by its original elements, but when they start being replaced
by "X" and "O", it stops creating newlines where they should be. The
board is displayed deformed because of this. To fix it, I used the
board array's indexes to create newlines instead.