complete play algorithm.
There were many bugs with rounds ending
prematurely that had to be fixed. It was fixed by
ensuring the method which checks for winning
combos returns false unless it detects a win. The
bug came from that method returning true even
when there wasn't a win, so it would satisfy the
condition to end the game, once the minimum #
of rounds to win was achieved.